templater: parse map {key => value} construct
Similar to the lambda construct, the map literal cannot be used as a first-class
template expression. That's the main reason why I decided to add a literal
syntax instead of make_map(key=value, ..) function. A template function has to
return a property of a certain type, but the templater doesn't support generics
or heterogeneous containers. OTOH, a literal expression can be interpreted
whatever the surrounding function expects. A downside is that this makes the
language bigger.

Regarding the syntax, I originally considered copying it from Python, but the
":" separator might conflict with the <kind>:<expr> pattern construct. So I
decided to use "=>" as a separator. Since our template language is somewhat
similar to Rust and Ruby, the use of "=>" seems good. Another potential problem
is that { .. } can be ambiguous if we add block expression. This could be
avoided by using square brackets, but then, an empty [] becomes ambiguous if we
add array literal. I don't have strong feeling about [] vs {}, though.

https://github.com/jj-vcs/jj/issues/3262#issuecomment-2031595004