annotate: read (some) file contents concurrently across commits
With high-latency backends like the one at Google, reading the file
contents one version at time is very slow. This patch changes the
annotation code to materialize the tree value concurrently across
commits. That means we call `Backend::read_file()` for many versions,
but we won't start reading from the `AsyncRead` until later, so it
should be fine even if the file is large. For the Google backend, the
`Backend::read_file()` causes the backend to fetch the first chunk of
the file revision from the server. The chunks are large enough that we
typically don't need to fetch any more chunks (but small enough that
we won't run out of memory even if we have `Backend::concurrency()`
chunks in memory at once).

This speeds up `jj file annotate` at Google on a relatively small
sample file from ~90 s to ~5 s.