commit_builder: allow recreating commit, don't create cycle in parent mapping
When we stop including predecessors in the Commit object, we're going
to sometimes end up creating commits with the same id. Commit
f438ac2fb240 made that an error. However, I don't think we can really
prevent these cases, so we need to handle them better instead. This
patch therefore removes the error and instead makes sure that we don't
create direct cycles in the predecessor and parent mappings. There's
still a risk of creating longer cycles, but I don't think we have any
cases where that can happen, so this should be good enough in practice
for now at least.

We ran into this in one of our backends at Google. That backend uses
the Git hashing scheme, so it doesn't include the predecessors and it
has second-level timestamp resolution. It would therefore somewhat
frequently end up rewriting a commit with the same id as already
existed.