governance: Require that maintainers should strive for transparency and communicate their decisions
This was taken out of the discussion in #6966. It spells out the _minimal_ requirements 
to the maintainer group. However it is left open, how it will be done since that depends on 
the set of maintainers and their individual experiences and preferences. 

Thank you @nasamuffin for helping me out of the writers block here.