backend: assert that tree entries are sorted also in release build
We had a scare at Google in our Git-like commit backend yesterday
where we were building `Tree` objects with entries out of order
(because the inputs use the Git order). Luckily we never ended up
writing incorrectly sorted entries to the database, but I can easily
imagine that happening to someone else's backend in the future. This
patch therefore promotes the `debug_assert!` in
`Tree::from_sorted_entries()` to a real `assert!`. Having this
assertion in release builds will also make it easier to diagnose any
such bugs in the future.

It should be relatively cheap to check that entries are in order but I
haven't done any measurements.