docs: use pymdownx.blocks extensions for admonition, details, tab
In preparation for rendering the `jj help -k ...` output in a richer
format.

Currently, the entire content of the "Example" blocks (and tabs, etc),
when rendered by a typical markdown parser, shows up as a single code
block due to being indented. However, these are meant to be rendered
as normal markdown, just inside a special container when rendered by
mkdocs.

Using the newer `pymdownx.blocks` extensions changes the syntax such
that using standard markdown viewers renders the content correctly.