rustfmt.toml: don't format generated files
This mostly affects the files in `lib/src/protos`, and makes it so that
`rustfmt` invoked via `jj fix --include-unchanged-files` does not format
those.

This avoids errors like
https://github.com/jj-vcs/jj/actions/runs/16586832790/job/46913734860?pr=7117

An interesting fact is that this commit does *not* affect `cargo fmt`.
`cargo fmt` skips those files either way because they are not proper
modules, but are rather `include!`-d in their `mod.rs`.