cli: evolog: add --no-operation flag to disable printing operation
I have a script that parses the output of `jj evolog` in order to find
the most recent version of a change on a remote. However, it's been
broken since the change that added `-- operation xyz` display. The
script could be fixed with some `sed` magic to strip out the unused
information, but I think it's generally to be able to disable the
information.

In lieu of adding a `CommitAndOperation` template type, this adds a
flag similar to the `--no-patch` flag to `jj show`. Show the operation
information by default, elide if asked.